Yolo Object detection sample modified
Implementation is a slight modification of the OV Yolo sample to demo a simple and basic person/bottle detection along with a virtual basket logic for billing.
eg:
If a person picks up a bottle, it is added to the virtual basket and the person is billed. If the bottle is put back on the shelf, the  price of the bottle is deducted from the person's bill.